backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.26.0
structlog==24.1.0
prometheus-client==0.19.0
tenacity==8.2.3
python-dateutil==2.8.2
email-validator==2.1.0.post1
